--- src/main/java/fdse/zc/fix/symbol/Fix.java
@@ -1,7 +1,0 @@
-package fdse.zc.fix.symbol;
-public class Fix {
-  public static void main(String[] args) throws Exception {
-
-  }
-
-}
